feat: remove rust-web-core
