--- crates/rust-wasm-core/Cargo.toml
@@ -1,21 +1,0 @@
-[package]
-name = "rust-wasm-core"
-version = "0.1.0"
-authors = ["michaelbguo <michaelbguo@tencent.com>"]
-edition = "2021"
-
-[lib]
-crate-type = ["cdylib", "rlib"]
-
-[features]
-default = ["console_error_panic_hook"]
-
-[dependencies]
-console_error_panic_hook = { version = "^0.1.7", optional = true }
-wee_alloc = { version = "^0.4.5", optional = true }
-wasm-bindgen = "^0.2"
-js-sys = "^0.3"
-lazy_static = "^1.4"
-
-[dev-dependencies]
-wasm-bindgen-test = "^0.3"
