--- crates/rust-wasm-core/package.json
@@ -1,11 +1,0 @@
-{
-  "name": "rust-wasm-core",
-  "version": "1.0.0",
-  "scripts": {
-    "build": "wasm-pack build --release",
-    "test": "cargo test"
-  },
-  "devDependencies": {
-    "wasm-pack": "^0.10.2"
-  }
-}
