+++ crates/rust-wasm-web/Cargo.toml
@@ -11,11 +11,12 @@ crate-type = ["cdylib", "rlib"]
 default = ["console_error_panic_hook"]
 
 [dependencies]
-console_error_panic_hook = { version = "^0.1.6", optional = true }
+console_error_panic_hook = { version = "^0.1.7", optional = true }
 wee_alloc = { version = "^0.4.5", optional = true }
-wasm-bindgen = "^0.2.63"
-lazy_static = "^1.4.0"
 oxipng = "^5.0.1"
+wasm-bindgen = "^0.2"
+js-sys = "^0.3"
+lazy_static = "^1.4"
 
 [dev-dependencies]
 wasm-bindgen-test = "^0.3.13"
