+++ crates/rust-wasm-web/src/adapter.rs
