+++ crates/rust-wasm-web/src/api.rs
@@ -1,17 +1,6 @@
-#[macro_use]
-extern crate lazy_static;
-
+use crate::encoder::safe_encode;
 use js_sys::*;
 use wasm_bindgen::prelude::*;
-
-use crate::utils::safe_encode;
-
-// 全局常量
-lazy_static! {
-    static ref CALL_COUNT: u32 = 100;
-}
-
-mod utils;
 
 // 浏览器平台全局函数
 #[wasm_bindgen]
@@ -34,27 +23,22 @@ extern "C" {
     fn console_log(s: &str);
 }
 
-#[wasm_bindgen]
 pub fn parse_number(params: &Number) -> JsString {
     safe_encode(&format!("?number={}", params))
 }
 
-#[wasm_bindgen]
 pub fn parse_int(params: &BigInt) -> JsString {
     safe_encode(&format!("{}", params))
 }
 
-#[wasm_bindgen]
 pub fn parse_str(params: &str) -> JsString {
     safe_encode(&params)
 }
 
-#[wasm_bindgen]
 pub fn parse_string(params: String) -> JsString {
     safe_encode(&params)
 }
 
-#[wasm_bindgen]
 pub fn parse_boolean(params: Boolean) -> JsString {
     if params == true {
         return safe_encode("true");
@@ -62,17 +46,14 @@ pub fn parse_boolean(params: Boolean) -> JsString {
     safe_encode("false")
 }
 
-#[wasm_bindgen]
 pub fn parse_null(params: &JsValue) -> bool {
     JsValue::is_null(params)
 }
 
-#[wasm_bindgen]
 pub fn parse_undefined(params: &JsValue) -> bool {
     JsValue::is_undefined(params)
 }
 
-#[wasm_bindgen]
 pub fn parse_symbol(params: &Symbol) -> String {
     let mut result: String = "Not a Symbol".to_owned();
     if JsValue::is_symbol(&params) {
@@ -81,7 +62,6 @@ pub fn parse_symbol(params: &Symbol) -> String {
     result
 }
 
-#[wasm_bindgen]
 pub fn parse_object(params: &Object) -> Object {
     let result = Object::new();
     if JsValue::is_null(params) {
@@ -90,7 +70,6 @@ pub fn parse_object(params: &Object) -> Object {
     result
 }
 
-#[wasm_bindgen]
 pub fn parse_set(params: &Set) -> Box<[JsValue]> {
     let mut arr: Vec<JsValue> = Vec::new();
     params.for_each(&mut |el, _, _| {
@@ -99,7 +78,6 @@ pub fn parse_set(params: &Set) -> Box<[JsValue]> {
     arr.into_boxed_slice()
 }
 
-#[wasm_bindgen]
 pub fn reverse_map(params: &Map) -> Map {
     let result = Map::new();
     params.for_each(&mut |v, k| {
@@ -114,7 +92,6 @@ pub fn reverse_map(params: &Map) -> Map {
 }
 
 // 数组过滤bigint
-#[wasm_bindgen]
 pub fn parse_bigint_array(params: Option<Box<[JsValue]>>) -> Box<[JsValue]> {
     let mut result: Vec<JsValue> = Vec::new();
     if let Some(arr) = params {
@@ -128,20 +105,16 @@ pub fn parse_bigint_array(params: Option<Box<[JsValue]>>) -> Box<[JsValue]> {
 }
 
 // ArrayBuffer = [u8]
-#[wasm_bindgen]
 pub fn parse_array_buffer(params: &[u8]) -> usize {
     params.len()
 }
 
-#[wasm_bindgen]
-pub fn call_js_method() -> u32 {
+pub fn call_js_method() {
     takes_immutable_closure(&|| {
         console_log("call_js_method in rust!");
     });
-    CALL_COUNT.to_owned()
 }
 
-#[wasm_bindgen]
 pub fn call_take_method(params: &str) {
     // 可变函数 利用闭包与rust交互
     takes_closure_parse_to_string(&|x| format!("{}={}&ticket=F08300", params, x));
