+++ package.json
@@ -13,6 +13,7 @@
     "lint": "turbo run lint",
     "clear": "turbo run clear",
     "tsnode": "ts-node-esm scripts/sw.ts",
+    "update": "turbo run update --parallel",
     "format": "prettier --loglevel=warn --write \"**/*.{mjs,ts,tsx,scss,md}\""
   },
   "type": "module",
@@ -37,7 +38,7 @@
     "@rollup/plugin-wasm": "^5.2.0",
     "@rollup/pluginutils": "^4.2.1",
     "@swc/cli": "^0.1.57",
-    "@swc/core": "^1.2.223",
+    "@swc/core": "^1.2.224",
     "@swc/jest": "^0.2.22",
     "@types/canvas-size": "^1.2.0",
     "@types/express": "^4.17.13",
@@ -45,8 +46,8 @@
     "@types/pngjs": "^6.0.1",
     "@types/react": "^17.0.48",
     "@types/react-dom": "^17.0.17",
-    "@typescript-eslint/eslint-plugin": "^5.32.0",
-    "@typescript-eslint/parser": "^5.32.0",
+    "@typescript-eslint/eslint-plugin": "^5.33.0",
+    "@typescript-eslint/parser": "^5.33.0",
     "css-loader": "^6.7.1",
     "css-minimizer-webpack-plugin": "^3.4.1",
     "eslint": "^7.32.0",
@@ -62,13 +63,13 @@
     "imagemin-svgo": "^10.0.1",
     "imagemin-webp": "^7.0.0",
     "mini-css-extract-plugin": "^2.6.1",
-    "postcss": "^8.4.14",
+    "postcss": "^8.4.16",
     "postcss-loader": "^6.2.1",
     "postcss-preset-env": "^6.7.1",
     "prettier": "^2.7.1",
     "rimraf": "^3.0.2",
     "rollup": "^2.77.2",
-    "sass": "^1.54.3",
+    "sass": "^1.54.4",
     "sass-loader": "^12.6.0",
     "sharp": "^0.30.7",
     "style-loader": "^3.3.1",
